add readme and fix linters
