+++ README.md
@@ -1,2 +1,102 @@
-# ToDo-List
-An app to create tasks for each project, build with JavaScript 
+
+<p align="center">
+  <a href="https://github.com/jcy2704/oop-ruby">
+    <img src="https://res.cloudinary.com/growsurf-prod/image/upload/v1582211139/production/gnysw2objzekbagrqiax.png" alt="Logo" width="350" height="70">
+  </a>
+</p>
+
+<h1 align="center">TO-DO LIST</h1>
+
+<p align="center">
+  <strong>Made by Jenny Cárdenas</strong>
+  <br>
+   This is a single page application for a to-do list.
+
+</p>
+
+<p align="center">
+  <a href="https://github.com/janis-jenny/ToDo-List/issues">
+    <img src="https://img.shields.io/badge/REPORT%20A%20BUG-purple?style=for-the-badge">
+  </a>
+   ‎ ‎ ‎ ‎
+  <a href="https://github.com/janis-jenny/ToDo-List/issues">
+    <img src="https://img.shields.io/badge/Request%20a%20feature-purple?style=for-the-badge">
+  </a>
+</p>
+
+
+
+## 👩🏼‍💻 About To-Do List
+
+![screenshot]()
+
+All page information dynamically generated and component loading achieved using pure Javascript. You can include your to-dos on different projects, edit its details and delete the to-do if you completed it or the whole project.Persistence is with local storage.
+
+<hr>
+
+
+## 🔴 Live Demo:
+
+[Demo](https://raw.githack.com/janis-jenny/ToDo-List/todoApp/dist/index.html)
+
+
+
+## 🔧 Built With
+
+- HTML5
+- CSS3
+- JavaScript ES6
+- Webpack
+- Linters
+- VS Code
+
+
+## 🤖 How it works
+
+To get a local copy up and running follow these simple steps:
+
+1. Go to the [repository page](https://github.com/janis-jenny/Restaurant-Page/).
+2. Press the "Code" button and copy the link.
+3. Clone it using git command `git clone <you_repository_link>`.
+4. Move to the local project: cd `<Your_repository_name>`.
+5. Run `npm install` to get all the dependencies.
+6. If you want to run the development mode run: `npm run dev`.
+7. Run `npm run watch` to run webpack build in watch mode.
+
+
+
+## 👥 Author
+
+👤 **Jenny Cárdenas**
+
+- GitHub: [@janis_jenny](https://github.com/janis-jenny)
+- Twitter: [@janis_jenny](https://twitter.com/janis_jenny)
+- LinkedIn: [Jenny Càrdenas](https://www.linkedin.com/in/paolajenny)
+
+
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](https://github.com/janis-jenny/ToDo-List/issues).
+
+
+
+## ⭐ Show your support
+
+Give a ⭐️ if you like this project!
+
+
+
+## 📌 Acknowledgments
+
+- [Stack Overflow](https://es.stackoverflow.com/)
+- [Microverse](https://www.microverse.org/)
+- [Odin Project](https://www.theodinproject.com/paths/full-stack-javascript/courses/javascript/lessons/todo-list)
+- Stand up team
+
+
+## 📝 License
+
+This project is [MIT](https://opensource.org/licenses/MIT) licensed.
