+++ src/components/common.js
@@ -17,33 +17,18 @@ const getValue = () => {
 
 const getIndex = (id) => storage.findIndex((project) => project.id === id);
 
-const displayTodoForm = (element) => element.classList.remove('d-none')
+const displayTodoForm = (element) => element.classList.remove('d-none');
 
-const hideTodoForm = (element) => element.classList.add('d-none')
+const hideTodoForm = (element) => element.classList.add('d-none');
 
 const cancelButton = (element) => {
   const cancel = document.querySelector('.cancel-button');
-  cancel.addEventListener('click' , e => {
+  cancel.addEventListener('click', e => {
     e.preventDefault();
     element.classList.toggle('d-none');
   });
-}
-
-const setLocalStorage = () => {
-  window.localStorage.setItem('todos', JSON.stringify(storage));
-};
-
-
-
-const getLocalStorage = () => {
-  const colletion = JSON.parse(window.localStorage.getItem('todos'));
-  if (colletion) {
-    colletion.forEach(el => {
-      storage.push(el);
-    });
-  }
 };
 
 export {
-  populateSelect, getValue, setLocalStorage, getLocalStorage, getIndex, storage, hideTodoForm, displayTodoForm, cancelButton
+  populateSelect, getValue, getIndex, storage, hideTodoForm, displayTodoForm, cancelButton,
 };
