+++ src/index.js
@@ -1,6 +1,6 @@
 import { populateSelect } from './components/common';
 import {
-  createProject, createTodo, displayProjects, displayTasks, showTaskForm, showProjectForm
+  createProject, createTodo, displayProjects, displayTasks, showTaskForm, showProjectForm,
 } from './components/render';
 import Project from './components/project';
 import './css/stylesheet.css';
